# src/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: AIRIS EPM API
  description: |
    AIRIS Enterprise Performance Management (EPM) API provides comprehensive monitoring, 
    analytics, and performance optimization capabilities for enterprise systems.
    
    ## Features
    - Real-time performance monitoring
    - Data quality validation
    - Advanced analytics and reporting
    - Caching and optimization
    - Alert and notification system
    
    ## Authentication
    API uses Bearer token authentication. Include your API key in the Authorization header.
    
    ## Rate Limiting
    API requests are limited to 1000 requests per hour per API key.
    
  version: "1.0.0"
  contact:
    name: AIRIS EPM Support
    url: https://airis-epm.com/support
    email: support@airis-epm.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://api.airis-epm.com/v1
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /alerts:
    get:
      tags:
        - Alerts
      summary: Get alerts
      description: Retrieve system alerts
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, resolved, acknowledged]
          description: Alert status filter
        - name: severity
          in: query
          schema:
            type: string
            enum: [low, medium, high, critical]
          description: Alert severity filter
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
          description: Maximum number of alerts to return
      responses:
        '200':
          description: Alerts retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertsResponse'

    post:
      tags:
        - Alerts
      summary: Create alert rule
      description: Create a new alert rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlertRuleRequest'
      responses:
        '201':
          description: Alert rule created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertRuleResponse'
  /alerts/{alertId}:
    get:
      tags:
        - Alerts
      summary: Get specific alert
      description: Retrieve details of a specific alert
      parameters:
        - name: alertId
          in: path
          required: true
          schema:
            type: string
          description: Alert identifier
      responses:
        '200':
          description: Alert retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertDetails'
        '404':
          description: Alert not found

    patch:
      tags:
        - Alerts
      summary: Update alert status
      description: Update the status of an alert
      parameters:
        - name: alertId
          in: path
          required: true
          schema:
            type: string
          description: Alert identifier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlertStatusUpdate'
      responses:
        '200':
          description: Alert status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertDetails'

  # Health Check
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AlertsResponse:
      type: object
      properties:
        success:
          type: boolean
        alerts:
          type: array
          items:
            $ref: '#/components/schemas/Alert'
        total_count:
          type: integer
    Alert:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        severity:
          type: string
          enum: [low, medium, high, critical]
        status:
          type: string
          enum: [active, resolved, acknowledged]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        rule_id:
          type: string
    AlertDetails:
      allOf:
        - $ref: '#/components/schemas/Alert'
        - type: object
          properties:
            history:
              type: array
              items:
                $ref: '#/components/schemas/AlertEvent'
            related_metrics:
              type: array
              items:
                type: string
    AlertEvent:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        event_type:
          type: string
          enum: [created, triggered, acknowledged, resolved]
        message:
          type: string
        user:
          type: string
    AlertRuleRequest:
      type: object
      required:
        - name
        - condition
        - severity
      properties:
        name:
          type: string
        description:
          type: string
        condition:
          type: string
        severity:
          type: string
          enum: [low, medium, high, critical]
        enabled:
          type: boolean
          default: true
        notification_channels:
          type: array
          items:
            type: string
    AlertRuleResponse:
      type: object
      properties:
        success:
          type: boolean
        rule:
          $ref: '#/components/schemas/AlertRule'
    AlertRule:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        condition:
          type: string
        severity:
          type: string
          enum: [low, medium, high, critical]
        enabled:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AlertStatusUpdate:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum: [acknowledged, resolved]
        comment:
          type: string
...